requirements.txt
# torch>=1.8
tensorboard
numpy
scipy
open3d>=0.8
Pillow
tqdm
MinkowskiEngine